Buggy Pause Menu, you can pause and unpause the game with the P key or by pressing the resume button Problems: Pressing the button that takes you back to the main menu in the pause menu does not update something, so the player and enemy are still drawn and playable on the main menu screen.  Somewhere the state of the player and enemy are not being reset when the gamestate is changed to the main menu screen
Pausing and unpusing the game speeds everything up even though the speed valuse are set at 600 for the player speed.  the only thing that I can find that changes is the shootingTime variable, so whatever is increasing that value is also somehow increasing the speed of everything else.  Might be solved by calling an update to reset variable values.